demo/python/edge/requirements.txt
opentelemetry-api>=1.31.1,<2.0.0
lambda-otel-lite>=0.12.0,<1.0.0
opentelemetry-sdk-extension-aws>=2.1.0 
opentelemetry-propagator-aws-xray>=1.0.2,<2.0.0